Scheduler ignore region for testing shoots
