+++ pkg/operation/botanist/health_check.go
@@ -405,6 +405,7 @@ func (b *HealthChecker) CheckClusterNodes(
 	condition gardencorev1beta1.Condition,
 	nodeLister kutil.NodeLister,
 ) (*gardencorev1beta1.Condition, error) {
+
 	for _, worker := range workers {
 		requirement, err := labels.NewRequirement("worker.gardener.cloud/pool", selection.Equals, []string{worker.Name})
 		if err != nil {
@@ -427,15 +428,21 @@ func (b *HealthChecker) CheckClusterNodes(
 			return &c, nil
 		}
 	}
+
 	return nil, nil
 }
 
 // CheckMonitoringSystemComponents checks whether the monitoring in the given listers are complete and healthy.
 func (b *HealthChecker) CheckMonitoringSystemComponents(
 	namespace string,
+	isTestingShoot bool,
 	condition gardencorev1beta1.Condition,
 	daemonSetLister kutil.DaemonSetLister,
 ) (*gardencorev1beta1.Condition, error) {
+
+	if isTestingShoot {
+		return nil, nil
+	}
 
 	daemonSetList, err := daemonSetLister.DaemonSets(namespace).List(monitoringSelector)
 	if err != nil {
@@ -448,17 +455,23 @@ func (b *HealthChecker) CheckMonitoringSystemComponents(
 	if exitCondition := b.checkDaemonSets(condition, daemonSetList); exitCondition != nil {
 		return exitCondition, nil
 	}
+
 	return nil, nil
 }
 
 // CheckMonitoringControlPlane checks whether the monitoring in the given listers are complete and healthy.
 func (b *HealthChecker) CheckMonitoringControlPlane(
 	namespace string,
+	isTestingShoot bool,
 	wantsAlertmanager bool,
 	condition gardencorev1beta1.Condition,
 	deploymentLister kutil.DeploymentLister,
 	statefulSetLister kutil.StatefulSetLister,
 ) (*gardencorev1beta1.Condition, error) {
+
+	if isTestingShoot {
+		return nil, nil
+	}
 
 	deploymentList, err := deploymentLister.Deployments(namespace).List(monitoringSelector)
 	if err != nil {
@@ -517,10 +530,15 @@ func (b *HealthChecker) CheckOptionalAddonsSystemComponents(
 // CheckLoggingControlPlane checks whether the logging components in the given listers are complete and healthy.
 func (b *HealthChecker) CheckLoggingControlPlane(
 	namespace string,
+	isTestingShoot bool,
 	condition gardencorev1beta1.Condition,
 	deploymentLister kutil.DeploymentLister,
 	statefulSetLister kutil.StatefulSetLister,
 ) (*gardencorev1beta1.Condition, error) {
+
+	if isTestingShoot {
+		return nil, nil
+	}
 
 	deploymentList, err := deploymentLister.Deployments(namespace).List(loggingSelector)
 	if err != nil {
@@ -573,11 +591,11 @@ func (b *Botanist) checkControlPlane(
 	if exitCondition, err := checker.CheckControlPlane(b.Shoot.Info, b.Shoot.SeedNamespace, condition, seedDeploymentLister, seedStatefulSetLister, seedWorkerLister); err != nil || exitCondition != nil {
 		return exitCondition, err
 	}
-	if exitCondition, err := checker.CheckMonitoringControlPlane(b.Shoot.SeedNamespace, b.Shoot.WantsAlertmanager, condition, seedDeploymentLister, seedStatefulSetLister); err != nil || exitCondition != nil {
+	if exitCondition, err := checker.CheckMonitoringControlPlane(b.Shoot.SeedNamespace, b.Shoot.GetPurpose() == gardencorev1beta1.ShootPurposeTesting, b.Shoot.WantsAlertmanager, condition, seedDeploymentLister, seedStatefulSetLister); err != nil || exitCondition != nil {
 		return exitCondition, err
 	}
 	if gardenletfeatures.FeatureGate.Enabled(features.Logging) {
-		if exitCondition, err := checker.CheckLoggingControlPlane(b.Shoot.SeedNamespace, condition, seedDeploymentLister, seedStatefulSetLister); err != nil || exitCondition != nil {
+		if exitCondition, err := checker.CheckLoggingControlPlane(b.Shoot.SeedNamespace, b.Shoot.GetPurpose() == gardencorev1beta1.ShootPurposeTesting, condition, seedDeploymentLister, seedStatefulSetLister); err != nil || exitCondition != nil {
 			return exitCondition, err
 		}
 	}
@@ -601,7 +619,7 @@ func (b *Botanist) checkSystemComponents(
 	if exitCondition, err := checker.CheckSystemComponents(b.Shoot.Info.Status.Gardener.Version, metav1.NamespaceSystem, condition, shootDeploymentLister, shootDaemonSetLister); err != nil || exitCondition != nil {
 		return exitCondition, err
 	}
-	if exitCondition, err := checker.CheckMonitoringSystemComponents(metav1.NamespaceSystem, condition, shootDaemonSetLister); err != nil || exitCondition != nil {
+	if exitCondition, err := checker.CheckMonitoringSystemComponents(metav1.NamespaceSystem, b.Shoot.GetPurpose() == gardencorev1beta1.ShootPurposeTesting, condition, shootDaemonSetLister); err != nil || exitCondition != nil {
 		return exitCondition, err
 	}
 	if exitCondition, err := checker.CheckOptionalAddonsSystemComponents(metav1.NamespaceSystem, condition, shootDeploymentLister, shootDaemonSetLister); err != nil || exitCondition != nil {
